WIP: add cli to tie types tracing and registry building from file
