+++ ahnlich/typegen/Cargo.toml
@@ -9,3 +9,4 @@ serde-reflection = "0.3.6"
 serde_json = "1.0.116"
 ndarray.workspace = true
 types = { path = "../types", version = "*" }
+clap.workspace = true
