+++ ahnlich/typegen/src/cli/mod.rs
@@ -1,0 +1,3 @@
+mod generate;
+
+pub use generate::{Cli, Commands};
